# Update and optimize LodePNG