… src/lodepng/lodepng_util.cpp