Backport and adapt newer functions with previous optimizations, further optimize some existing functions, update formatting

These changes decrease the size of the LodePNG compiled code and speed up optimizing PNGs using the --allfilters switch

On MSYS2 MINGW64 gcc 12.2.0, LodePNG code is 
 - 14.24% smaller (160682 bytes -> 137800 bytes, 22.35kb smaller)
 - ~10.57% faster using the command ect -4096 -strip --allfilters "[neat image.png](https://user-images.githubusercontent.com/97068837/187946543-2d8c8889-fe83-483a-9213-90a983f1322f.png)" (47.78114s -> 42.73037s)

On MSYS2 CLANG64 clang 14.0.6, LodePNG code is 
 - 12.06% smaller (192040 bytes -> 168972 bytes, 22.53kb smaller) 
 - ~10.97% faster using the command ect -4096 -strip --allfilters "[neat image.png](https://user-images.githubusercontent.com/97068837/187946543-2d8c8889-fe83-483a-9213-90a983f1322f.png)" (44.03142s -> 39.20114s)

There's a ~0.5% speed increase for the --allfilters-b switch, otherwise there is no measurable speed difference when optimizing PNGs without using --allfilters or --allfilters-b