@fhanau 2022-09-01
Thank you for the pull request. ECT's current lodepng is very difficult to maintain due to the hack and slash changes I made years ago before adopting proper code style, so bringing it closer to main lodepng can certainly be helpful. This is a whole lot of code to look at though, so I'll need a while to go through it.