@woot000 2022-09-01
I left some of the LodePNG functions unchanged (mostly some of the pixel decoding functions, an example being getPixelColorRGBA8) because I was having some issues with corrupted output, and I couldn't find the exact function that was responsible for those issues (it might be unfilterScanlines but I'm not sure)

You could probably also update those too for a small boost in decoding speed