@fhanau 2022-09-02
Overall this is a whole lot of changes for one commit, making it difficult to review. It would have been useful to split it up into several commits (e.g. refactor, update to current lodepng, several commits with changes...).

I will take a look and first update the code base to current lodepng (whole lot of changes already) and then try to identify useful changes in here going beyond that.