@fhanau 2022-09-08
I looked at the changes to mainline lodepng in the past few years and adopted many of them on the new lodepng-work branch. I plan to merge it after testing the changes, overall it should make little difference in terms of performance.

As far as I can see, your pull request includes useful changes for optimize_palette, filterScanline and filter (let me know if I'm missing something) on top of that, but also includes lots of code style changes that deviate from mainline lodepng. I'm trying to align this code more closely with it so that it is easier to maintain in the future, so some of the code style changes are not helpful.

Perhaps you can file smaller pull requests based on the new branch that reduce code size/improve performance, it should just be commits that are smaller so they are easier to review.