@woot000 2022-09-08
The optimizations are primarily in filter, filterScanline, and optimize_palette (I may be forgetting a few). I'll resubmit this pr with those changes in smaller bunches of commits soon